Move AlignSeries to align.go
